Move enum declaration into concern
